# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TrackStar API",
    "description": "Personal Task & Habit Tracker API",
    "version": "1.0.0",
    "contact": {
      "name": "TrackStar Team",
      "email": "support@trackstar.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/tasks/": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get all tasks for the current user",
        "description": "Retrieve all tasks belonging to the authenticated user, sorted by creation date.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a new task",
        "description": "Create a new task for the authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Task information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Learn Node.js"
                },
                "description": {
                  "type": "string",
                  "example": "Complete Node.js tutorial"
                },
                "priority": {
                  "type": "string",
                  "example": "high"
                },
                "dueDate": {
                  "type": "string",
                  "example": "2024-12-31"
                },
                "category": {
                  "type": "string",
                  "example": "learning"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get a specific task by ID",
        "description": "Retrieve details of a specific task by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Task ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Update an existing task",
        "description": "Update details of an existing task by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Task ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated task information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Updated task title"
                },
                "description": {
                  "type": "string",
                  "example": "Updated description"
                },
                "priority": {
                  "type": "string",
                  "example": "medium"
                },
                "isCompleted": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete a task",
        "description": "Delete a specific task by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Task ID"
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for authentication (temporary for testing)"
      }
    }
  }
}